feat(歌曲与歌单的绑定): :sparkles: 初步解决歌单与歌曲的绑定关系

+++ apps/music-api/music-core/music-infrastructure/music-infrastructure-persistence/src/main/java/io/gitee/pkmer/core/infrastructure/persistence/songlist/mybatis/ListSong.java
@@ -17,8 +17,8 @@ public class ListSong {
 
     private Long id;
 
-    private Integer songId;
+    private Long songId;
 
-    private Integer songListId;
+    private Long songListId;
 
 }
